Impl a test-migrations GitHub workflow to verify migrations (#33)
(1) A GitHub Workflow was implemented to verify that running the
migrations will lead to the desired schema.

(2) Note: Typically we've been using Postgres 15.2. However, in this
workflow Postgres 14.11 is used. We do this so that the `pg_dump`
command can work with the service. The command comes from
postgresql-client from ubuntu-latest. Currently this is for Postgres
14.11.

(3) Note: There are two jobs—(i) dump-schema and (ii)
dump-migrations—that are identical copies of each other, just
parameterised differently by a `MIGRATIONS_DIR` env variable. It would
be nice if this can be a template of sorts that gets called.