enable row level security (#36)
(1) Enable Row Level Security (RLS) on all tables^1 and views.

(2) Reorganise the pipeline into workflows that trigger on pull-request
or on push to main. Workflows for testing and migrating are moved into
x-test and x-migrate and run only when called upon.

(3) Updated the test to run postgres in a docker container so that we do
not need postgres-client. This way we are not limited to using postgres
14 and can use postgres 15 instead. We need postgres 15 because we need
to specify `WITH (security_invoker=on)` when defining views; and this
was made available only in postgres 15.

^1 - It is not possible to enable RLS on flyway_schema_history using
Flyway because the table is locked by flyway when migration are run. RLS
will need to be enabled on flyway_schema_history outside of flyway.